requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.48.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
gradio==5.0.0
gradio_client==0.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pandas==2.2.0
pillow==10.3.0
plotly==5.19.0
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.18
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.1
safetensors==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.40.0
streamlit==1.37.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
websockets==11.0.3
Werkzeug==3.0.6
zipp==3.17.0
